% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_canada_2022_electoral_ridings.R
\docType{data}
\name{names_canada_2022_electoral_ridings}
\alias{names_canada_2022_electoral_ridings}
\title{Names of Canadian Electoral Ridings (2022 Redistribution)}
\format{
A data frame with columns:
\describe{
\item{id_riding}{Character. The unique identifier for each riding.}
\item{name_riding_en}{Character. The name of the riding in English.}
\item{name_riding_fr}{Character. The name of the riding in French.}
\item{id_province}{Character. The identifier of the province to which the riding belongs.}
}
}
\source{
Statistics Canada
}
\usage{
names_canada_2022_electoral_ridings
}
\description{
This dataset contains the names and identifiers of Canadian electoral ridings based on the 2022 redistribution.
}
\keyword{datasets}
